- updated to 3.0.2 - removed obsolete link,rules patches - added pc patch - by default build with SFST and openfst enabled, foma disabled
Changed files:
    hfst-link.patch -> 1.2
    hfst-pc.patch -> 1.1
    hfst-rules.patch -> 1.2
    hfst.spec -> 1.2

+++ hfst.spec
@@ -1,21 +1,27 @@
+#
+# Conditional build:
+%bcond_with	foma	# use foma by linking with libfoma (GPL v2-strict, which is not compliant)
+#
 Summary:	Helsinki Finite-State Transducer (library and application suite)
 Summary(pl.UTF-8):	Helsinki Finite-State Transducer (biblioteka i zestaw aplikacji)
 Name:		hfst
-Version:	2.4.1
+Version:	3.0.2
 Release:	1
-License:	GPL v3+
+License:	GPL v3
 Group:		Applications/Text
 Source0:	http://downloads.sourceforge.net/hfst/%{name}-%{version}.tar.gz
-# Source0-md5:	4e501d68f1ef67ab45029ce14e3ce559
-Patch0:		%{name}-link.patch
-Patch1:		%{name}-rules.patch
+# Source0-md5:	ae502571684f706b372d669c36652892
+Patch0:		%{name}-pc.patch
 URL:		http://www.ling.helsinki.fi/kieliteknologia/tutkimus/hfst/
-BuildRequires:	autoconf >= 2.61
-BuildRequires:	automake
+BuildRequires:	SFST-devel
+BuildRequires:	autoconf >= 2.62
+BuildRequires:	automake >= 1:1.11
 BuildRequires:	bison
+%{?with_foma:BuildRequires:	foma-devel}
 BuildRequires:	flex >= 2.5.35
 BuildRequires:	libstdc++-devel
 BuildRequires:	libtool >= 2:2.0
+BuildRequires:	openfst-devel
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 
 %description
@@ -34,7 +40,10 @@ Summary:	Header files for HFST library
 Summary(pl.UTF-8):	Pliki nagłówkowe biblioteki HFST
 Group:		Development/Libraries
 Requires:	%{name} = %{version}-%{release}
+Requires:	SFST-devel
+%{?with_foma:Requires:	foma-devel}
 Requires:	libstdc++-devel
+Requires:	openfst-devel
 
 %description devel
 Header files for HFST library.
@@ -57,38 +66,16 @@ Statyczna biblioteka HFST.
 %prep
 %setup -q
 %patch0 -p1
-%patch1 -p1
 
 %build
-cd hfst2
 %{__libtoolize}
 %{__aclocal}
 %{__autoconf}
 %{__autoheader}
-%{__automake}
-cd ../hfst-tools
-%{__libtoolize}
-%{__aclocal}
-%{__autoconf}
-%{__autoheader}
-%{__automake}
-cd hfst-lexc
-%{__libtoolize}
-%{__aclocal}
-%{__autoconf}
-%{__autoheader}
-%{__automake}
-cd ../hfst-twolc
-%{__libtoolize}
-%{__aclocal}
-%{__autoconf}
-%{__autoheader}
-%{__automake}
-cd ../..
-%{__aclocal}
-%{__autoconf}
 %{__automake}
 %configure \
+	FOMACLI=/usr/bin/foma \
+	--disable-silent-rules \
 	--enable-static
 
 %{__make}
@@ -99,6 +86,13 @@ rm -rf $RPM_BUILD_ROOT
 %{__make} install \
 	DESTDIR=$RPM_BUILD_ROOT
 
+# missing in make install
+[ ! -f $RPM_BUILD_ROOT%{_pkgconfigdir}/hfst.pc ] || exit 1
+install -D libhfst/hfst.pc $RPM_BUILD_ROOT%{_pkgconfigdir}/hfst.pc
+
+# obsoleted by pkgconfig
+%{__rm} $RPM_BUILD_ROOT%{_libdir}/libhfst.la
+
 %clean
 rm -rf $RPM_BUILD_ROOT
 
@@ -107,34 +101,19 @@ rm -rf $RPM_BUILD_ROOT
 
 %files
 %defattr(644,root,root,755)
-%doc AUTHORS BUGS ChangeLog NEWS README THANKS
+%doc AUTHORS ChangeLog NEWS README THANKS
 %attr(755,root,root) %{_bindir}/hfst-*
 %attr(755,root,root) %{_bindir}/htwolcpre*
-%attr(755,root,root) %{_bindir}/hwfst-calculate
 %attr(755,root,root) %{_libdir}/libhfst.so.*.*.*
-%attr(755,root,root) %ghost %{_libdir}/libhfst.so.0
-%attr(755,root,root) %{_libdir}/libhfstlexc.so.*.*.*
-%attr(755,root,root) %ghost %{_libdir}/libhfstlexc.so.0
+%attr(755,root,root) %ghost %{_libdir}/libhfst.so.1
 %{_mandir}/man1/hfst-*.1*
 
 %files devel
 %defattr(644,root,root,755)
 %attr(755,root,root) %{_libdir}/libhfst.so
-%attr(755,root,root) %{_libdir}/libhfstlexc.so
-%{_libdir}/libhfst.la
-%{_libdir}/libhfstlexc.la
-%{_includedir}/hfst2
-%{_includedir}/flex-utils.h
-%{_includedir}/lexc.h
-%{_includedir}/lexcio.h
-%{_includedir}/string-munging.h
-%{_includedir}/xducer.h
-%{_includedir}/xymbol.h
-%{_includedir}/xymbol-bridges.h
-%{_includedir}/xymbol-table.h
+%{_includedir}/hfst
 %{_pkgconfigdir}/hfst.pc
 
 %files static
 %defattr(644,root,root,755)
 %{_libdir}/libhfst.a
-%{_libdir}/libhfstlexc.a
